feat(build): 添加MSVC编译器运行时库动态链接设置
为了解决Windows平台下静态链接导致的内存管理问题，在CMakeLists.txt中添加了
MSVC编译器的运行时库动态链接配置(/MD和/MDd)，确保所有模块使用相同的
运行时库实例。

fix(examples): 重构示例文件命名并更新API调用

将examples目录下的示例文件重新命名以简化名称，并更新API调用方式，
包括使用AutoConnect替代Connect，移除不必要的设备信息获取代码，
并添加OpenTactile和CloseTactile调用。

refactor(dexhand): 优化MoveJoints函数实现

修改MoveJoints函数参数名从commands到joints，并使用unordered_map
进行关节ID索引优化。按照JointId枚举顺序遍历关节，确保硬件期望
的固定顺序数据传输。

refactor(types): 将JointId改为强类型枚举

将JointId从普通枚举改为class类型的强类型枚举，提高类型安全性，
避免隐式转换错误。

fix(core): 修复多个变量初始化和数值计算问题

修复ClearFault和InitJoint函数中result变量未正确初始化的问题，
修复温度变化检测中的浮点数比较问题，以及文件锁操作中的长度
类型转换问题。

+++ src/dexhand_callback_manager.cc
@@ -79,7 +79,7 @@ bool DexHandCallbackManager::HasTemperatureChanged(const HandTemperature& temper
     }
 
     // 2. 温度变化触发（>1°C）
-    float temp_diff = std::abs(temperature.temperature - last_temperature_.temperature);
+    int temp_diff = std::abs(temperature.temperature - last_temperature_.temperature);
     if (temp_diff > TEMPERATURE_THRESHOLD) {
         return true;
     }
